requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
livekit-api
livekit-agents
livekit-plugins-cartesia
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-turn-detector
openai
redis
httpx
openpyxl
weasyprint
